Added All links Html

+++ src/FrontEnd.php
@@ -24,10 +24,12 @@ class FrontEnd {
      * Initialize hook after  plugins_loaded
      */
     public static function init() {
+
         self::getOption();
         self::setFilename();
 
         if (isset(self::$option->enable) && !isset(self::$option->disable_loggedin)) {
+            //var_dump(self::$fileName);
             if (file_exists(self::$fileName)) {
                 //CacheLoader::run();//plugins_loaded Event
                 //echo "Init Event";
@@ -38,12 +40,18 @@ class FrontEnd {
                 header("pragma:public");
                 echo file_get_contents(self::$fileName);
                 if (is_user_logged_in()) {
-                    echo self::cacheTools();
+                    self::cacheTools();
                 }
                 self::die();
             } else {
                 self::FrontEndInit();
             }
+        } else {
+            if (is_user_logged_in()) {
+                add_action('wp_footer', [self::class, 'cacheTools']);
+                //echo self::cacheTools();
+            }
+            self::FrontEndInit();
         }
     }
 
@@ -110,16 +118,18 @@ class FrontEnd {
         </style>
         <div class="cache-toolbar" title="Cache Toolbar">
             <div class='tolbar-head'><svg xmlns="http://www.w3.org/2000/svg" style="max-width:22px" class="svg-icon" viewBox="0 0 512 512"><title>Cache Tools</title><path d="M315.27 33L96 304h128l-31.51 173.23a2.36 2.36 0 002.33 2.77h0a2.36 2.36 0 001.89-.95L416 208H288l31.66-173.25a2.45 2.45 0 00-2.44-2.75h0a2.42 2.42 0 00-1.95 1z" fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="28"/></svg></div>
+
             <a href="javascript:void(0)" onclick="refreshCache(this)"><svg style="max-width:22px" xmlns="http://www.w3.org/2000/svg" class="ionicon" viewBox="0 0 512 512"><title>Refresh Cache</title><path d="M256 48C141.31 48 48 141.32 48 256c0 114.86 93.14 208 208 208 114.69 0 208-93.31 208-208 0-114.87-93.13-208-208-208zm0 313a94 94 0 010-188h4.21l-14.11-14.1a14 14 0 0119.8-19.8l40 40a14 14 0 010 19.8l-40 40a14 14 0 01-19.8-19.8l18-18c-2.38-.1-5.1-.1-8.1-.1a66 66 0 1066 66 14 14 0 0128 0 94.11 94.11 0 01-94 94z"/></svg></a>
             <a href="javascript:void(0)" id="cahceRemove" onclick="removeCache(this)"><svg style="max-width:22px" xmlns="http://www.w3.org/2000/svg" class="ionicon" viewBox="0 0 512 512"><title>Remove Cache</title><path d="M432 448a15.92 15.92 0 01-11.31-4.69l-352-352a16 16 0 0122.62-22.62l352 352A16 16 0 01432 448zM431.5 204a16 16 0 00-15.5-12H307.19L335.4 37.63c.05-.3.1-.59.13-.89A18.45 18.45 0 00302.73 23l-92.58 114.46a4 4 0 00.29 5.35l151 151a4 4 0 005.94-.31l60.8-75.16A16.37 16.37 0 00431.5 204zM301.57 369.19l-151-151a4 4 0 00-5.93.31L83.8 293.64A16.37 16.37 0 0080.5 308 16 16 0 0096 320h108.83l-28.09 154.36v.11a18.37 18.37 0 0032.5 14.53l92.61-114.46a4 4 0 00-.28-5.35z"/></svg></a>
         </div>
         <script>
-            let doneSvg = '<svg xmlns="http://www.w3.org/2000/svg" style="max-width:22px" class="ionicon" viewBox="0 0 512 512"><title>Process Done</title><path style="color:#3af93a" fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="32" d="M416 128L192 384l-96-96"/></svg>';
-            let loadingSvg = '<svg xmlns="http://www.w3.org/2000/svg" style="max-width:22px" class="loading-cahce-process" viewBox="0 0 512 512"><title>Processing</title><path d="M434.67 285.59v-29.8c0-98.73-80.24-178.79-179.2-178.79a179 179 0 00-140.14 67.36m-38.53 82v29.8C76.8 355 157 435 256 435a180.45 180.45 0 00140-66.92" fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="32"/><path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="32" d="M32 256l44-44 46 44M480 256l-44 44-46-44"/></svg>';
-            let hideTime;
-            let inProcess = false;
-            let ajaxurl = '<?php echo admin_url('admin-ajax.php') ?>';
-            let CurrentUrl = window.location.href;
+            var doneSvg = '<svg xmlns="http://www.w3.org/2000/svg" style="max-width:22px" class="ionicon" viewBox="0 0 512 512"><title>Process Done</title><path style="color:#3af93a" fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="32" d="M416 128L192 384l-96-96"/></svg>';
+            var loadingSvg = '<svg xmlns="http://www.w3.org/2000/svg" style="max-width:22px" class="loading-cahce-process" viewBox="0 0 512 512"><title>Processing</title><path d="M434.67 285.59v-29.8c0-98.73-80.24-178.79-179.2-178.79a179 179 0 00-140.14 67.36m-38.53 82v29.8C76.8 355 157 435 256 435a180.45 180.45 0 00140-66.92" fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="32"/><path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="32" d="M32 256l44-44 46 44M480 256l-44 44-46-44"/></svg>';
+            var hideTime;
+            var inProcess = false;
+            var ajaxurl = '<?php echo admin_url('admin-ajax.php') ?>';
+            var CurrentUrl = window.location.href;
+
             jQuery(window).mousemove(function () {
                 jQuery('.cache-toolbar').removeClass('cache-toolbar-hide');
                 clearTimeout(hideTime);
@@ -163,7 +173,7 @@ class FrontEnd {
             }
         </script>
         <?php
-        return ob_get_clean();
+        echo ob_get_clean();
     }
 
     /**
@@ -187,22 +197,53 @@ class FrontEnd {
      * @return type
      */
     static function storeCache($html) {
+        if (self::is_ajax()) {
+            return $html;
+        }
         //createCache::getSuccCachedInfo();
+        if (is_user_logged_in()) {
+            //echo "<pre>";
+            //var_dump($_SERVER);
+            //echo "</pre>";
+            $actual_link = (isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on' ? "https" : "http") . "://$_SERVER[HTTP_HOST]$_SERVER[REQUEST_URI]";
+            $WPHttp = new \WP_Http_Curl();
+            $arg = [
+                'headers' => [
+                    'User-Agent' => isset($_SERVER['HTTP_USER_AGENT']) ? $_SERVER['HTTP_USER_AGENT'] : ""
+                ],
+                'stream' => false,
+                'filename' => "",
+                'decompress' => false
+            ];
+            $WPHttp->request($actual_link, $arg);
+        } else {
 
-        $headers = headers_list();
-        if (self::$option->response_header == "") {
-            self::$option->response_header = implode("\n", $headers);
+            self::setFilename();
+            if (file_exists(self::$fileName)) {//If Exist then skip 
+                return $html;
+            }
+            $headers = headers_list();
+            if (self::$option->response_header == "") {
+                self::$option->response_header = implode("\n", $headers);
+                self::updateOption((array) self::$option);
+            }
+            //$html;
+            self::dirInit();
+            if (trim($html) != "") {
+                self::storeInfo();
+                file_put_contents(self::$fileName, $html);
+            }
+        }
 
-            self::updateOption((array) self::$option);
-        }
-        //$html;
-        self::dirInit();
-        if (trim($html) != "") {
-            self::storeInfo();
-            file_put_contents(self::$fileName, $html);
-        }
         //createCache::putSuccCachedInfo();
         return $html;
+    }
+
+    public static function is_ajax() {
+        if (!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest') {
+            return true;
+        }
+        return false;
     }
 
     public static function storeInfo() {
